Refactor Google Map integration
